backend/pyproject.toml
[tool.poetry]
name = "haqnow-community"
version = "0.1.0"
description = "Community document platform backend"
authors = ["Haqnow Team <ops@haqnow.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
sqlalchemy = "^2.0.32"
pyjwt = "^2.9.0"
passlib = {version = "^1.7.4", extras = ["bcrypt"]}
pyotp = "^2.9.0"
boto3 = "^1.34.162"
python-dotenv = "^1.0.1"
celery = "^5.3.4"
redis = "^5.0.1"
pillow = "^10.4.0"
pymupdf = "^1.24.10"
pytesseract = "^0.3.13"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
httpx = "^0.27.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^5.0.0"
black = "^24.0.0"
isort = "^5.13.0"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]


